+++ tests/TestCase/TestSuite/ConsoleIntegrationTestCaseTest.php
@@ -14,6 +14,7 @@ namespace MeTools\Test\TestCase\TestSuite;
 
 use App\Shell\ChildExampleShell;
 use App\Shell\ExampleShell;
+use Cake\TestSuite\Stub\ConsoleOutput;
 use MeTools\TestSuite\ConsoleIntegrationTestCase;
 
 /**
@@ -22,36 +23,20 @@ use MeTools\TestSuite\ConsoleIntegrationTestCase;
 class ConsoleIntegrationTestCaseTest extends ConsoleIntegrationTestCase
 {
     /**
+     * @var Cake\TestSuite\Stub\ConsoleOutput
+     */
+    protected $_out;
+
+    /**
      * Called before every test method
      * @return void
      */
     public function setUp()
     {
         $this->Shell = $this->getMockForShell(ExampleShell::class);
+        $this->_out = new ConsoleOutput;
 
         parent::setUp();
-    }
-
-    /**
-     * Test for `assertTableHeadersEquals()` method
-     * @test
-     */
-    public function testAssertTableHeadersEquals()
-    {
-        $this->exec('example print_table');
-        $class = $this->Shell;
-        $this->assertTableHeadersEquals($class::$tableHeaders);
-    }
-
-    /**
-     * Test for `assertTableRowsEquals()` method
-     * @test
-     */
-    public function testAssertTableRowsEquals()
-    {
-        $this->exec('example print_table');
-        $class = $this->Shell;
-        $this->assertTableRowsEquals($class::$tableRows);
     }
 
     /**
@@ -60,10 +45,31 @@ class ConsoleIntegrationTestCaseTest extends ConsoleIntegrationTestCase
      */
     public function testGetShellMethods()
     {
-        $this->assertEquals(['doNothing', 'printTable'], $this->getShellMethods());
-        $this->assertEquals(['printTable'], $this->getShellMethods(['doNothing']));
+        $this->assertEquals(['aSimpleMethod', 'doNothing'], $this->getShellMethods());
+        $this->assertEquals(['aSimpleMethod'], $this->getShellMethods(['doNothing']));
 
         $this->Shell = $this->getMockForShell(ChildExampleShell::class);
-        $this->assertEquals(['childMethod', 'doNothing', 'printTable'], $this->getShellMethods());
+        $this->assertEquals(['aSimpleMethod', 'childMethod', 'doNothing'], $this->getShellMethods());
+    }
+
+    /**
+     * Test for `assertOutputNotEmpty()` method
+     * @test
+     */
+    public function testAssertOutputNotEmpty()
+    {
+        $this->_out->write('message');
+        $this->assertOutputNotEmpty();
+    }
+
+    /**
+     * Test for `assertOutputNotEmpty()` method, on failure
+     * @expectedException PHPUnit\Framework\ExpectationFailedException
+     * @expectedExceptionMessage stdout was empty
+     * @test
+     */
+    public function testAssertOutputNotEmptyOnFailure()
+    {
+        $this->assertOutputNotEmpty();
     }
 }